requirements.txt
instagrapi==2.1.3
python-dotenv==1.0.1
schedule==1.2.2
rich==13.9.4
click==8.1.8
SQLAlchemy==2.0.36
Pillow==11.0.0
httpx==0.27.2
